requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cleanup==6.0.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
phonenumberslite==8.12.56
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
